refactor Break out script Add ability to mock scanner input through command args
Add requirements.txt
Add .gitignore

Update Dockerfile
