+++ Dockerfile
@@ -1,7 +1,7 @@
-FROM python
+FROM python:3
 
-RUN pip install evdev
+RUN pip install -r requirements.txt
 
-ADD script.py /
+ADD src/ /script
 
-ENTRYPOINT python /script.py
+ENTRYPOINT python3 /script/script.py
