+++ requirements.txt
@@ -1,0 +1,1 @@
+evdev==1.7.0
