+++ src/util.py
@@ -1,0 +1,4 @@
+import os
+
+def envvar(key, default=None):
+    return os.environ.get(key, default)
